rf: move `DandisetLinkmlTranslationReport`
From `cli/models.py` to `models.py`
