+++ src/dandisets_linkml_status_tools/models.py
@@ -1,5 +1,12 @@
+from datetime import datetime
 from typing import Any, NamedTuple
 from collections.abc import Sequence
+
+from dandi.dandiapi import VersionStatus
+from dandisets_linkml_status_tools.cli.models import (
+    DandisetMetadataType,
+    LinkmlValidationErrsType,
+)
 
 from typing_extensions import TypedDict  # Required for Python < 3.12 by Pydantic
 
@@ -160,3 +167,49 @@ class JsonschemaValidationErrorType(NamedTuple):
             and type(self.validator_value) is type(other.validator_value)  # noqa E721
             and self.validator_value == other.validator_value
         )
+
+
+class DandisetLinkmlTranslationReport(BaseModel):
+    """
+    A report of validation results of a dandiset (metadata) against the
+    `dandischema.models.Dandiset` or `dandischema.models.PublishedDandiset` Pydantic
+    model and the corresponding LinkML schema.
+    """
+
+    dandiset_identifier: str
+    dandiset_version: str  # The version of the dandiset being validated
+
+    @property
+    def dandiset_schema_version(self) -> str:
+        """
+        The schema version of the dandiset metadata being validated as specified in the
+        metadata itself
+
+        :return: The schema version of the dandiset metadata being validated. An empty
+            indicates that no valid schema version is found in the metadata.
+        """
+        version = self.dandiset_metadata.get("schemaVersion", "")
+
+        # Since there is no guarantee that the dandiset metadata is valid,
+        # there is no guarantee the obtained version is a valid string.
+        # If the fetched version is not a string, return an empty string to indicate,
+        # there is no valid version.
+        if not isinstance(version, str):
+            version = ""
+
+        return version
+
+    # Dandiset version status as provided by the DANDI API
+    dandiset_version_status: VersionStatus
+
+    # Dandiset version modified datetime as provided by the DANDI API
+    dandiset_version_modified: datetime
+
+    # The metadata of the dandiset to be validated
+    dandiset_metadata: DandisetMetadataType
+
+    # Error encountered in validation against the Pydantic dandiset metadata model
+    pydantic_validation_errs: Json[PydanticValidationErrsType] = []
+
+    # Errors encountered in validation against the dandiset metadata model in LinkML
+    linkml_validation_errs: LinkmlValidationErrsType = []
